- Nukes misplaced bmime package in d. Mime operations now in util/file - Removes misplaced VerifyCensorshipRecordFiles function from d/api/v1 to util/file - Mime/file operations now centralized in util/file as single source of truth
