+++ politeiad/backend/gitbe/gitbe.go
@@ -30,7 +30,6 @@ import (
 	"github.com/decred/politeia/decredplugin"
 	pd "github.com/decred/politeia/politeiad/api/v1"
 	"github.com/decred/politeia/politeiad/api/v1/identity"
-	"github.com/decred/politeia/politeiad/api/v1/mime"
 	"github.com/decred/politeia/politeiad/backend"
 	"github.com/decred/politeia/util"
 	filesystem "github.com/otiai10/copy"
@@ -398,7 +397,7 @@ func verifyContent(metadata []backend.MetadataStream, files []backend.File, file
 		f.digest = dp
 
 		// Verify MIME
-		detectedMIMEType := mime.DetectMimeType(f.payload)
+		detectedMIMEType := util.DetectMimeType(f.payload)
 		if detectedMIMEType != files[i].MIME {
 			return nil, backend.ContentVerificationError{
 				ErrorCode: pd.ErrorStatusInvalidMIMEType,
@@ -409,7 +408,7 @@ func verifyContent(metadata []backend.MetadataStream, files []backend.File, file
 			}
 		}
 
-		if !mime.MimeValid(files[i].MIME) {
+		if !util.MimeValid(files[i].MIME) {
 			return nil, backend.ContentVerificationError{
 				ErrorCode: pd.ErrorStatusUnsupportedMIMEType,
 				ErrorContext: []string{
