+++ politeiawww/cmd/politeiawwwcli/commands/editproposal.go
@@ -8,7 +8,6 @@ import (
 	"io/ioutil"
 	"path/filepath"
 
-	"github.com/decred/politeia/politeiad/api/v1/mime"
 	"github.com/decred/politeia/politeiawww/api/v1"
 	"github.com/decred/politeia/util"
 )
@@ -125,7 +124,7 @@ func (cmd *EditProposalCmd) Execute(args []string) error {
 
 	f := v1.File{
 		Name:    "index.md",
-		MIME:    mime.DetectMimeType(md),
+		MIME:    util.DetectMimeType(md),
 		Digest:  hex.EncodeToString(util.Digest(md)),
 		Payload: base64.StdEncoding.EncodeToString(md),
 	}
@@ -142,7 +141,7 @@ func (cmd *EditProposalCmd) Execute(args []string) error {
 
 		f := v1.File{
 			Name:    filepath.Base(file),
-			MIME:    mime.DetectMimeType(attachment),
+			MIME:    util.DetectMimeType(attachment),
 			Digest:  hex.EncodeToString(util.Digest(attachment)),
 			Payload: base64.StdEncoding.EncodeToString(attachment),
 		}
