+++ util/file.go
@@ -4,32 +4,98 @@ import (
 	"crypto/sha256"
 	"encoding/base64"
 	"encoding/hex"
+	"errors"
 	"io"
 	"io/ioutil"
+	"net/http"
 	"os"
 	"path/filepath"
 	"strings"
 
-	"github.com/decred/politeia/politeiad/api/v1/mime"
+	svg "github.com/h2non/go-is-svg"
+
+	"github.com/decred/dcrtime/merkle"
+	"github.com/decred/politeia/politeiad/api/v1"
+	"github.com/decred/politeia/politeiad/api/v1/identity"
 )
 
-// MimeFile returns the MIME type of a file.
-func MimeFile(filename string) (string, error) {
-	f, err := os.Open(filename)
-	if err != nil {
-		return "", err
-	}
-	defer f.Close()
+var (
+	// validMimeTypesMap is a list of all acceptable MIME types that
+	// can be communicated between client and server, structured
+	// as a map for fast access.
+	validMimeTypesMap = make(map[string]struct{})
 
-	// We need up to 512 bytes
-	b := make([]byte, 512)
-	n, err := f.Read(b)
-	if err != nil {
-		return "", err
+	ErrUnsupportedMimeType = errors.New("unsupported MIME type")
+)
+
+// MimeValid returns true if the passed string is a valid
+// MIME type, false otherwise.
+func MimeValid(s string) bool {
+	_, ok := validMimeTypesMap[s]
+	return ok
+}
+
+// DetectMimeType returns the file MIME type
+func DetectMimeType(data []byte) string {
+	// svg needs a specific check because the algorithm
+	// implemented by http.DetectContentType doesn't detect svg
+	if svg.IsSVG(data) {
+		return "image/svg+xml"
+	}
+	return http.DetectContentType(data)
+}
+
+// SetMimeTypesMap sets valid mimetypes list with loaded config
+func SetMimeTypesMap(validMimeTypesCfg []string) {
+	for _, m := range validMimeTypesCfg {
+		validMimeTypesMap[m] = struct{}{}
+	}
+}
+
+// Verify ensures that a CensorshipRecord properly describes the array of
+// files.
+func VerifyCensorshipRecordFiles(pid identity.PublicIdentity, csr v1.CensorshipRecord, files []v1.File) error {
+	digests := make([]*[sha256.Size]byte, 0, len(files))
+	for _, file := range files {
+		payload, err := base64.StdEncoding.DecodeString(file.Payload)
+		if err != nil {
+			return v1.ErrInvalidBase64
+		}
+
+		// MIME
+		mimeType := DetectMimeType(payload)
+		if !MimeValid(mimeType) {
+			return ErrUnsupportedMimeType
+		}
+
+		// Digest
+		h := sha256.New()
+		h.Write(payload)
+		d := h.Sum(nil)
+		var digest [sha256.Size]byte
+		copy(digest[:], d)
+
+		digests = append(digests, &digest)
 	}
 
-	// Clip buffer to prevent detecting binary files.
-	return mime.DetectMimeType(b[:n]), nil
+	// Verify merkle root
+	root := merkle.Root(digests)
+	if hex.EncodeToString(root[:]) != csr.Merkle {
+		return v1.ErrInvalidMerkle
+	}
+
+	s, err := hex.DecodeString(csr.Signature)
+	if err != nil {
+		return v1.ErrInvalidHex
+	}
+	var signature [identity.SignatureSize]byte
+	copy(signature[:], s)
+	r := hex.EncodeToString(root[:])
+	if !pid.VerifyMessage([]byte(r+csr.Token), signature) {
+		return v1.ErrCorrupt
+	}
+
+	return nil
 }
 
 // DigestFile returns the SHA256 of a file.
@@ -77,9 +143,9 @@ func LoadFile(filename string) (mimeType string, digest string, payload string, 
 	}
 
 	// MIME
-	mimeType = mime.DetectMimeType(b)
-	if !mime.MimeValid(mimeType) {
-		err = mime.ErrUnsupportedMimeType
+	mimeType = DetectMimeType(b)
+	if !MimeValid(mimeType) {
+		err = ErrUnsupportedMimeType
 		return
 	}
